+++ Source/DotSpatial.Symbology/Properties/AssemblyInfo.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Reflection;
+using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
@@ -17,6 +18,8 @@ using System.Runtime.InteropServices;
 // to COM components.  If you need to access a type in this assembly from
 // COM, set the ComVisible attribute to true on that type.
 [assembly: ComVisible(false)]
+
+[assembly: InternalsVisibleTo("DotSpatial.Symbology.Tests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100c3dd1f197a92d762c304983777a9d64f99025d201245f953d72bc02183d857d0c198819823df8ea8982925e5b34b9cfa5b8cd046ced832be2c4b7290fdfd40ce3b1065fc7fb9463f11d9dd54559dc02014126c6d66c03822e35b2bbb14b71f3cf1abd785bb8cc0beafc9b02a73bfd355f208d9de4d9362e80818d4635f3c41c0")]
 
 // The following GUID is for the ID of the typelib if this project is exposed to COM
 [assembly: Guid("6756d9d3-46bb-4ff0-9103-0bfb28fd619b")]